// src/wire.d.ts
export interface components {
  schemas: {
// ...
    RemoteCreate: {
      name: components["schemas"]["LanguageText"];
      icon?: components["schemas"]["IconIdentifier"];
      description?: components["schemas"]["LanguageText"];
    } & (
      | {
          clone_from: components["schemas"]["EntityId"];
        }
      | {
          codeset_id: components["schemas"]["SimpleId"];
        }
      | {
          custom_codeset: {
            /** @default custom */
            manufacturer_id: string;
            device_name: string;
            device_type?: components["schemas"]["DeviceType"];
          };
        }
      | {
          kind: components["schemas"]["RemoteKind"];
          bt: components["schemas"]["BtRemoteCreateOptions"];
        }
    );
// ...
    RemoteUpdate: {
      name?: components["schemas"]["LanguageText"];
      icon?: components["schemas"]["IconIdentifier"];
      description?: components["schemas"]["LanguageText"];
      options?: {
        ir?: {
          /** @description Delay in milliseconds between sending IR commands, if a command contains multiple IR codes. */
          cmd_delay?: number;
          /**
           * @description Repeat each IR command in the dataset n times. Defaults to 0 (no repeat) if not specified.
           *     This setting is intended mainly for PRONTO codes with certain devices requiring the same command being
           *     sent twice (e.g. Sony and Epson devices).
           */
          repeat?: number;
          codeset?: {
            id: string;
          };
          output?: {
            device_id: string;
            port_id: string;
          };
        };
      };
    };
// ...
    ProfileRequest: {
      profile_id?: components["schemas"]["SimpleId"];
      name: components["schemas"]["Name"];
      icon?: components["schemas"]["IconIdentifier"];
      /** @description Create a restricted profile which cannot change settings. Switching profiles requires the admin pin. */
      restricted?: boolean;
      description?: components["schemas"]["Description"];
    };
// ...
  };
// ...
}